Check classes rewritten
