--- autotest-web/src/main/java/steps/AdministrationPageCheckSteps.java
@@ -1,27 +1,0 @@
-package steps;
-
-import actions.WebChecks;
-import com.codeborne.selenide.SelenideElement;
-import io.cucumber.java.ru.Если;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import ru.lanit.at.web.pagecontext.PageManager;
-
-public class AdministrationPageCheckSteps {
-
-    private PageManager pageManager;
-    private static final Logger LOGGER = LoggerFactory.getLogger(WebActionSteps.class);
-
-    public AdministrationPageCheckSteps(PageManager pageManager) {
-        this.pageManager = pageManager;
-    }
-
-    @Если("на текущей странице {string} {string} присутствует")
-    public void textAppearOnThePage(String elementName, String text) {
-        SelenideElement element = pageManager
-                .getCurrentPage()
-                .getElement(elementName);
-        WebChecks.elementTextEqualsExpectedText(element, text, null);
-        LOGGER.info("в инфоблоке '{}' присутствует текст '{}'", pageManager.getCurrentPage().name(), element.getText());
-    }
-}
