--- autotest-web/src/test/resources/features/LinkEmployee.feature
@@ -1,19 +1,0 @@
-#language:ru
-@linkemployee
-Функционал: Проверка страницы администрирования
-
-  Сценарий: Страница администрирования, проверка работоспособности ссылки 'Сотрудники'
-
-    Дано открыть 'http://178.154.246.238:58082/'
-    Пусть чекбокс 'Я желаю войти с админскими правами' отображается и не выбран
-    Тогда нажать на чекбокс 'Я желаю войти с админскими правами'
-    Пусть поле ввода 'логин' отображается
-    Тогда заполнить поле 'логин' значением 'hr'
-    Пусть поле ввода 'пароль' отображается
-    Тогда заполнить поле 'пароль' значением 'hrmhrm12345'
-    Пусть поле ввода 'токен' отображается
-    Тогда с текущими 'hr' и 'hrmhrm12345' получить токен
-    Пусть кнопка 'Войти' отображается
-    Тогда нажать на кнопку 'Войти'
-    Затем нажать на ссылку с текстом 'Сотрудники'
-    Если на текущей странице 'Заголовок таблицы' 'Изменить Сотрудник' присутствует
